Allow intermediate input files
After the global operations, terminated by the first input image,
imagemagick handles arguments with the logic of a sequence. Every
additional file argument pushes a new image to the sequence while every
operation is applied to the existing sequence. This can be interleaved,
i.e. we may add some images after operations. These non-global
operations are only applied to the preceding images:

    magick-script rose: rose: rose: rose: -resize '50x50!' rose: output.gif

Creates a gif of four frames of the original image in the size 50x50,
followed by one frame which is cropped off since it has the inherent
70x46 size of the special `rose` example image.